Added support for nanosleep.
The old usleep() has the granularity of 1 millisec because the sleep
queue is only traversed in the timer interrupt handler and the timer
is only programmed by the scheduler.

We want some high resolution sleep. The way to do it is to set the timer
to wake up before a thread goes to sleep. It is a little bit harder than
that because sometimes we can't program the timer if it has already been
programmed to fire before our wakeup time.

Some tricks have been done. See my comment in sched/sleep.c nanosleep() for
details.