Refine MinnowMax i2c driver.
Still not sure if it works correctly. The problem lies in the message condition
when doing read. From what I read on the Internet, the first write cmd doesn't need
to issue a STOP and the second read cmd should issue a RESTART but still not
STOP. A STOP should only be issued after all the data to be read has been transferred.
But I don't know who's gonna issue the STOP.